% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\docType{methods}
\name{assay,BQSummarizedExperiment,missing-method}
\alias{assay,BQSummarizedExperiment,missing-method}
\title{extract assay data}
\usage{
\S4method{assay}{BQSummarizedExperiment,missing}(x, i, ...)
}
\arguments{
\item{x}{BQSummarizedExperiment instance}

\item{i}{index for retrieval, ignored at present}

\item{\dots}{not used}
}
\description{
extract assay data
}
\note{
Very experimental approach to retrieving numerical data given
a SummarizedExperiment 'shell'.  We need more checking of
consistency between assay and clinical data before creating the
shell.  We use dcast to transform query result to a matrix, and
some 'individuals' may have multiple contributions ... we use
\code{fun.aggregate = max} and will see warnings until this is cleared up.
}
