% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\docType{methods}
\name{assayNames,BQSummarizedExperiment-method}
\alias{assayNames,BQSummarizedExperiment-method}
\title{get assay names}
\usage{
\S4method{assayNames}{BQSummarizedExperiment}(x, ...)
}
\description{
get assay names
}
