% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\docType{methods}
\name{assays,RESTfulSummarizedExperiment-method}
\alias{assays,RESTfulSummarizedExperiment-method}
\title{assays access for RESTfulSummarizedExperiment}
\usage{
\S4method{assays}{RESTfulSummarizedExperiment}(x, ..., withDimnames = TRUE)
}
\arguments{
\item{x}{instance of RESTfulSummarizedExperiment}

\item{\dots}{not used}

\item{withDimnames}{logical defaults to TRUE}
}
\description{
assays access for RESTfulSummarizedExperiment
}
