% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{banoSEMeta}
\alias{banoSEMeta}
\title{banoSEMeta: metadata RangedSummarizedExperiment shell for banovichSE}
\format{RangedSummarizedExperiment instance}
\source{
10.1371/journal.pgen.1004663
}
\usage{
banoSEMeta
}
\description{
banoSEMeta: metadata RangedSummarizedExperiment shell for banovichSE
}
\examples{
data(banoSEMeta)
banoSEMeta
}
\keyword{datasets}
