% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancan_SE.R
\name{basic_clinvars}
\alias{basic_clinvars}
\title{provide a collection of basic clinical variables to limit size of clinical data return}
\usage{
basic_clinvars()
}
\value{
character(20) vector of variable names
}
\description{
provide a collection of basic clinical variables to limit size of clinical data return
}
\note{
Use pancan_app() to determine other variables available
}
