% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\name{BQ3_Array}
\alias{BQ3_Array}
\title{create BQ3_Array instance given url (filepath) and entity (host) name}
\usage{
BQ3_Array(filepath)
}
\arguments{
\item{filepath}{a BQ3_Source instance}
}
\value{
an instance of \code{\link[DelayedArray]{DelayedArray-class}}
}
\description{
create BQ3_Array instance given url (filepath) and entity (host) name
}
\examples{
if (interactive()) {
  con = cgcConn("TCGA_hg38_data_v0")
  ss = BQ3_Source(con, filterval="TCGA-LUAD")
  BQ3_Array(ss)
}
}
