% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\name{BQ3_Source}
\alias{BQ3_Source}
\title{construct a BigQuery resource interface}
\usage{
BQ3_Source(bqconn, tblnm = "RNAseq_Gene_Expression",
  rowkeyfield = "Ensembl_gene_id", colkeyfield = "case_barcode",
  filtervbl = "project_short_name", filterval = "TCGA-GBM",
  assayvbl = "HTSeq__Counts")
}
\arguments{
\item{bqconn}{instance of BigQueryConnection from bigrquery}

\item{tblnm}{character(1) table name known to bqconn}

\item{rowkeyfield}{character(1) field in the table that will
generate rownames for matrices derived from table}

\item{colkeyfield}{character(1) field in the table that will
generate colnames for matrices derived from table}

\item{filtervbl}{character(1) field in the table that will be used to filter out a group of records,
for example, all records pertaining to a given tumor in TCGA}

\item{filterval}{character(1) value in the range of filtervbl to identify records to retain --}

\item{assayvbl}{character(1) field with assay quantifications}
}
\value{
instance of BQ3_Source
}
\description{
construct a BigQuery resource interface
}
\examples{
if (interactive()) {
 con = cgcConn("TCGA_hg38_data_v0")
 lungConn = BQ3_Source(con, filterval="TCGA-LUAD")
 lungConn
}
}
