% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQM_Array.R
\name{bqConn}
\alias{bqConn}
\title{generate a connection to BigQuery for specific dataset}
\usage{
bqConn(dataset, project, billing)
}
\arguments{
\item{dataset}{character(1) name of dataset in project}

\item{project}{character(1) name of project}

\item{billing}{character(1) billing code for project}
}
\value{
an instance of BigQueryConnection
}
\description{
generate a connection to BigQuery for specific dataset
}
\note{
You will need to authenticate with Google.
}
\examples{
bqConn
}
