% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQM_Array.R
\name{BQM_Array}
\alias{BQM_Array}
\title{create BQM_Array instance given url (filepath) and entity (host) name}
\usage{
BQM_Array(filepath)
}
\arguments{
\item{filepath}{a BQM_Source instance}
}
\value{
an instance of \code{\link[DelayedArray]{DelayedArray-class}}
}
\description{
create BQM_Array instance given url (filepath) and entity (host) name
}
\examples{

# authentication issues may arise.  if you are authorized
# to use bigquery with GPC project isb-cgc, a token may
# be generated through the following
# options(httr_oob_default=TRUE)
# example(BQM_Source)
# a browser authentication event may occur, or if you are in
# a browserless session, a URL will be emitted, possibly in
# the context of warnings ... browse to this URL and an
# authentication event will occur, and a token will be provided
# this can be provided back to the R session to allow the
# query to proceed
#
if (interactive()) {
  con = bqConn(project="cgc-05-0009", dataset="yriMulti", billing=
          Sys.getenv("CGC_BILLING"))
  ss = BQM_Source(con, "banovichSE_methylationData", "cg_Methyl450")
  BQM_Array(ss)
}
}
