% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQM_Array.R
\docType{class}
\name{BQM_Source-class}
\alias{BQM_Source-class}
\title{Represent information about a BigQuery resource with a 'triple' database schema.
This is targeting the isb-cgc TCGA layout.
BigQuery Records are regarded as triples, within major groups defined by filtervbl.
Triples have content subject - gene - value, to be pivoted to genes(rows) x 
subjects(columns) with values as entries.}
\description{
Represent information about a BigQuery resource with a 'triple' database schema.
This is targeting the isb-cgc TCGA layout.
BigQuery Records are regarded as triples, within major groups defined by filtervbl.
Triples have content subject - gene - value, to be pivoted to genes(rows) x 
subjects(columns) with values as entries.
}
