% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQM_Array.R
\name{BQM_Source}
\alias{BQM_Source}
\title{construct a BigQuery resource interface}
\usage{
BQM_Source(bqconn, tblnm, rowkeyfield, maxdfsize = 5e+05)
}
\arguments{
\item{bqconn}{instance of BigQueryConnection from bigrquery}

\item{tblnm}{character(1) table name known to bqconn}

\item{rowkeyfield}{character(1) field in the table that will
generate rownames for matrices derived from table}
}
\description{
construct a BigQuery resource interface
}
\examples{
if (interactive()) {
 con = bqConn("yriMulti")
 banoMeth = BQM_Source(con, tblnm="banovichSE_MethylationData",
     rowkeyfield="cg_methyl450")
 banoMeth
}
}
