% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\name{cgcConn}
\alias{cgcConn}
\title{simplify connection to a BigQuery dataset for ISB CGC}
\usage{
cgcConn(dataset = "TCGA_bioclin_v0", project = "isb-cgc",
  billing = Sys.getenv("CGC_BILLING"))
}
\arguments{
\item{dataset}{character string with dataset name}

\item{project}{character string with project name}

\item{billing}{string with billing code}
}
\description{
simplify connection to a BigQuery dataset for ISB CGC
}
