% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{datasetRefs}
\alias{datasetRefs}
\title{obtain dataset metadata from a set of links}
\usage{
datasetRefs(linkset, index, drop = 5)
}
\arguments{
\item{linkset}{instance of H5S_linkset}

\item{index}{numerical index of group to use}

\item{drop}{numerical indices of hosts to be ignored}
}
\description{
obtain dataset metadata from a set of links
}
