% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{datasets}
\alias{datasets}
\title{extract dataset UUID and attributes from H5ShContent}
\usage{
datasets(object, ...)
}
\arguments{
\item{object}{instance of H5ShContent}

\item{\dots}{not used yet}
}
\description{
extract dataset UUID and attributes from H5ShContent
}
