% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\name{DelayedArray,BQ3_ArraySeed-method}
\alias{DelayedArray,BQ3_ArraySeed-method}
\title{constructor}
\usage{
\S4method{DelayedArray}{BQ3_ArraySeed}(seed)
}
\arguments{
\item{seed}{instance of 'array'}
}
\description{
constructor
}
