% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\docType{methods}
\name{dim,BQ3_ArraySeed-method}
\alias{dim,BQ3_ArraySeed-method}
\title{dim derived from saved dimnames}
\usage{
\S4method{dim}{BQ3_ArraySeed}(x)
}
\arguments{
\item{x}{instance of BQ3_ArraySeed}
}
\value{
integer(2) vector of dimensions corresponding to R's layout, assuming 2-d data
}
\description{
dim derived from saved dimnames
}
