% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\docType{methods}
\name{dimnames,BQ3_ArraySeed-method}
\alias{dimnames,BQ3_ArraySeed-method}
\title{dimnames are saved in the BQ3_ArraySeed}
\usage{
\S4method{dimnames}{BQ3_ArraySeed}(x)
}
\arguments{
\item{x}{instance of BQ3_ArraySeed}
}
\value{
currently returns list(NULL, NULL) as we do not store dimnames in HDF5
}
\description{
dimnames are saved in the BQ3_ArraySeed
}
