% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQM_Array.R
\docType{methods}
\name{dimnames,BQM_ArraySeed-method}
\alias{dimnames,BQM_ArraySeed-method}
\title{dimnames are saved in the BQM_ArraySeed}
\usage{
\S4method{dimnames}{BQM_ArraySeed}(x)
}
\arguments{
\item{x}{instance of BQM_ArraySeed}
}
\value{
currently returns list(NULL, NULL) as we do not store dimnames in HDF5
}
\description{
dimnames are saved in the BQM_ArraySeed
}
