% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteArray.R
\docType{methods}
\name{dimnames,H5S_ArraySeed-method}
\alias{dimnames,H5S_ArraySeed-method}
\title{dimnames not stored with H5S_source as of Jan 2018}
\usage{
\S4method{dimnames}{H5S_ArraySeed}(x)
}
\arguments{
\item{x}{instance of H5S_ArraySeed}
}
\description{
dimnames not stored with H5S_source as of Jan 2018
}
