% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{dsmeta}
\alias{dsmeta}
\title{list information about datasets available in an H5S_source}
\usage{
dsmeta(src)
}
\arguments{
\item{src}{H5S_source instance}
}
\description{
list information about datasets available in an H5S_source
}
