% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_1Mneurons}
\alias{full_1Mneurons}
\title{full 1.3 million cell dataset from 10x genomics}
\format{RangedSummarizedExperiment instance}
\source{
\url{https://community.10xgenomics.com/t5/10x-Blog/Our-1-3-million-single-cell-dataset-is-ready-to-download/ba-p/276} massaged into RangedSummarizedExperiment shell
}
\usage{
full_1Mneurons
}
\description{
full 1.3 million cell dataset from 10x genomics
}
\examples{
data(full_1Mneurons)
full_1Mneurons
}
\keyword{datasets}
