% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{getH5ShContent}
\alias{getH5ShContent}
\title{acquire a list of REST content for an HDF5 server query, wrap in H5ShContent ('host' content)}
\usage{
getH5ShContent(serverURL, host)
}
\arguments{
\item{serverURL}{a character string giving the URL including port number on which HDF5 server is listening}

\item{host}{character}
}
\description{
acquire a list of REST content for an HDF5 server query, wrap in H5ShContent ('host' content)
}
\examples{
# this is data served by HDF group
tall = getH5ShContent( serverURL = "https://data.hdfgroup.org:7258", host="tall.data.hdfgroup.org")
tall
tds = datasets(tall)
tds
tds["0:3", "0:3"]
tds[c(1,3),]
tds[,c(1,3)]
tds[c(1,3),c(1,3)]
}
