% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{goPatt}
\alias{goPatt}
\title{Create a data.frame with ENSEMBL and SYMBOL identifiers associated with a GO TERM specified by a regular expression in \code{termPattern}}
\usage{
goPatt(termPattern = "neurotro", targets = c("ENSEMBL", "SYMBOL"),
  organism = "Hs", inst = "eg")
}
\arguments{
\item{termPattern}{a character string encoding a regular expression to be matched to keys of type TERM in GO.db}

\item{targets}{\code{columns} to be returned from org.[organism].[inst].db}

\item{organism}{two-letter code for organism in the OrgDb family of packages}

\item{inst}{two- or three-letter code (e.g., \code{eg} for ENTREZ GENE or \code{sgd} for yeastgenome.org) identifying institute responsible for annotation}
}
\value{
data.frame
}
\description{
Create a data.frame with ENSEMBL and SYMBOL identifiers associated with a GO TERM specified by a regular expression in \code{termPattern}
}
\examples{
gp = goPatt()
dim(gp)
head(gp)
}
