% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr450k}
\alias{gr450k}
\title{gr450k: GRanges with metadata for illumina 450k methylation assay}
\format{RangedSummarizedExperiment instance}
\source{
Bioconductor
}
\usage{
gr450k
}
\description{
gr450k: GRanges with metadata for illumina 450k methylation assay
}
\keyword{datasets}
