% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{groups}
\alias{groups}
\alias{groups,H5S_source,missing-method}
\title{HDF5 server data groups accessor}
\usage{
groups(object, index, ...)
}
\arguments{
\item{object}{H5S_source instance}

\item{index}{numeric, if present, extracts metadata about selected group (sequential ordering of groups as returned by server)
access for group information for HDF5 server}

\item{\dots}{not used}
}
\description{
HDF5 server data groups accessor
}
