% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{methods}
\name{groups,H5S_source,numeric-method}
\alias{groups,H5S_source,numeric-method}
\title{selective group metadata accessor}
\usage{
\S4method{groups}{H5S_source,numeric}(object, index, ...)
}
\arguments{
\item{object}{instance of H5S_source}

\item{index}{numeric}

\item{\dots}{unused}
}
\description{
selective group metadata accessor
}
