% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtexRecount}
\alias{gtexRecount}
\title{gtexRecount: metadata RangedSummarizedExperiment shell for RECOUNT gtex rse_gene}
\format{RangedSummarizedExperiment instance}
\source{
\url{http://duffel.rail.bio/recount/SRP059959/rse_gene.Rdata}
}
\usage{
gtexRecount
}
\description{
gtexRecount: metadata RangedSummarizedExperiment shell for RECOUNT gtex rse_gene
}
\keyword{datasets}
