% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5S_Array.R
\docType{class}
\name{H5S_Array-class}
\alias{H5S_Array-class}
\alias{as}
\alias{coerce,H5S_Array,H5S_Matrix-method}
\title{extension of DelayedArray for HDF Server content}
\description{
extension of DelayedArray for HDF Server content

coercion for remote array to remote matrix
}
