% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{class}
\name{H5S_source}
\alias{H5S_source}
\alias{H5S_source-class}
\alias{H5S_source}
\alias{[[,H5S_source,character,ANY-method}
\alias{dataset}
\title{H5S_source identifies an HDF5 server and manages some metadata about contents}
\usage{
H5S_source(serverURL, ...)

\S4method{[[}{H5S_source,character,ANY}(x, i, j)

dataset(h5s, tag)
}
\arguments{
\item{serverURL}{a URL for a port for HDF5Server}

\item{\dots}{not used}

\item{x}{instance of H5S_source}

\item{i}{character string intended to identify dataset on server}

\item{j}{not used}

\item{h5s}{instance of H5S_source}

\item{tag}{character string identifying a dataset}
}
\description{
H5S_source identifies an HDF5 server and manages some metadata about contents

construct H5S_source
}
\section{Slots}{

\describe{
\item{\code{serverURL}}{character string with a URL}

\item{\code{dsmeta}}{DataFrame instance with metadata about content of server}
}}

\note{
The dsmeta slot holds a DataFrame with a column \code{dsnames}
that is a list with ith element a character vector of all dsnames
available for the ith group.  There is no effort at present to
search all groups for candidate datasets.
}
\examples{
bigec2 = H5S_source("http://54.174.163.77:5000")
bigec2
dsmeta(bigec2)[1:2,] # two groups
dsmeta(bigec2)[1,2][[1]] # all dataset candidates in group 1
}
