% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{H5S_source}
\alias{H5S_source}
\title{construct H5S_source}
\usage{
H5S_source(serverURL, ...)
}
\arguments{
\item{serverURL}{a URL for a port for HDF5Server}

\item{\dots}{not used}
}
\description{
construct H5S_source
}
\examples{
bigec2 = H5S_source("http://54.174.163.77:5000")
bigec2
dsmeta(bigec2)
}
