% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{internalDim}
\alias{internalDim}
\title{acquire internal HDF5 dimension information for matrix}
\usage{
internalDim(h5d)
}
\arguments{
\item{h5d}{instance of H5S_dataset}
}
\description{
acquire internal HDF5 dimension information for matrix
}
