% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbCgc.R
\name{isbCgcDatasets}
\alias{isbCgcDatasets}
\title{Access the dataset names in the "isb-cgc" project}
\usage{
isbCgcDatasets()
}
\value{
character vector
}
\description{
Access the dataset names in the "isb-cgc" project
}
\examples{
isbCgcDatasets() # might be out of date ... can use list_datasets in bigrquery
}
