% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbCgc.R
\name{isbCgcTables}
\alias{isbCgcTables}
\title{list the tables in a selected dataset}
\usage{
isbCgcTables(dataset = "TCGA_hg19_data_v0",
  billing = Sys.getenv("CGC_BILLING"))
}
\arguments{
\item{dataset}{character string identifying a table in ISB CGC}

\item{billing}{Google BigQuery billing code, which can be set in an environment variable \code{CGC_BILLING}}
}
\description{
list the tables in a selected dataset
}
