% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BQ3_Array.R
\name{newDA}
\alias{newDA}
\title{DelayedArray bridge}
\usage{
newDA(seed = new("array"), Class = "DelayedArray")
}
\arguments{
\item{seed}{instance of 'array'}

\item{Class}{character(1)}
}
\description{
DelayedArray bridge
}
