% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteArray.R
\name{RemoteArray}
\alias{RemoteArray}
\title{create RemoteArray instance given url (filepath) and entity (host) name}
\usage{
RemoteArray(filepath, host)
}
\arguments{
\item{filepath}{a character(1) URL to port for HDF Server}

\item{host}{a character(1) name of 'host' in server}
}
\description{
create RemoteArray instance given url (filepath) and entity (host) name
}
\examples{
# The true values from yriMulti data element 'banovichSE':
# > assay(banovichSE[c(1:5,329465:329469),c(1:3,63:64)])
#                    NA18498    NA18499    NA18501  |    NA18489     NA18909
# cg00000029      0.47339629  1.2943041 -0.8084735  |  0.6708168 -0.86093022
# cg00000165      1.23640861  0.2099817 -0.2683763  |  0.4446088  0.99868231
# cg00000236     -0.22258183  1.6236857 -0.8654838  |  0.1958195 -0.06090929
# cg00000289      0.65720581  0.5527470 -1.8458295  | -0.4618782  0.34934164
# cg00000363     -0.15063083  0.7498020  0.3254333  |  0.7342878  0.12940774
# #-------------------------------------------------------------------------
# ch.9.98936572R -0.07954958  0.2139431 -0.4719621  |  0.6835012  0.57758798
# ch.9.98937537R  0.04254705  1.0702770  1.7356387  | -0.1531732 -1.52889773
# ch.9.98959675F -1.59253143  0.2982456 -1.1954030  | -1.3703135  0.28974909
# ch.9.98989607R -1.80646652  0.4760022  1.4771808  |  0.9479602  0.49921375
# ch.9.991104F    0.08180195 -0.2434306  1.0281002  | -0.1653721  0.55612215
#
# compare to that delivered by RemoteArray
#
RemoteArray("http://h5s.channingremotedata.org:5000", "assays")
}
