% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\docType{class}
\name{RESTfulSummarizedExperiment-class}
\alias{RESTfulSummarizedExperiment-class}
\title{hdf5server-based assay for SummarizedExperiment}
\description{
hdf5server-based assay for SummarizedExperiment
}
