% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\docType{methods}
\name{RESTfulSummarizedExperiment}
\alias{RESTfulSummarizedExperiment}
\alias{RESTfulSummarizedExperiment,RangedSummarizedExperiment,H5S_dataset-method}
\alias{.RESTfulSummarizedExperiment}
\alias{assay}
\alias{assay,RESTfulSummarizedExperiment,missing-method}
\title{Construct RESTfulSummarizedExperiment}
\usage{
RESTfulSummarizedExperiment(se, source)

.RESTfulSummarizedExperiment(se, source)

\S4method{assay}{RESTfulSummarizedExperiment,missing}(x, i, ...)
}
\arguments{
\item{se}{SummarizedExperiment instance, assay component can be empty SimpleList}

\item{source}{instance of H5S_dataset}

\item{x}{instance of RESTfulSummarizedExperiment}

\item{i}{not used}

\item{\dots}{not used}
}
\value{
instance of RESTfulSummarizedExperiment

matrix
}
\description{
Construct RESTfulSummarizedExperiment

hidden constructor
}
\note{
RESTfulSummarizedExperiment contains a global dimnames
list generated at creation.  It is possible that standard operations 
on a SummarizedExperiment will engender dimnames components that
differ from the initial global dimnames, principally through
uniqification (adding suffixes when dimname elements are
repeated).  When this is detected, assay() will fail with a complaint
about length(setdiff(*names(x), x@globalDimnames[[...]])).
}
\examples{
require("rhdf5client")
hsds = H5S_source(serverURL=URL_hsds())
hsdsCon = setPath(hsds,"/home/reshg/bano_meQTLex.h5")
fetchDatasets(hsdsCon)
banoh5 = H5S_dataset2(hsdsCon,"d-435d7ad4-9f13-11e8-92c2-0242ac120021")
ehub = ExperimentHub::ExperimentHub()
myfiles <- AnnotationHub::query(ehub , "restfulSEData")
myfiles[["EH551"]] -> banoSEMeta
rr = RESTfulSummarizedExperiment(banoSEMeta, banoh5)
rr
rr2 = rr[1:4, 1:5] # just modify metadata
rr2
assay(rr2) # extract data
}
