% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{se1.3M}
\alias{se1.3M}
\title{Convenience functions using HSDS server to extract tenx neurons full or subset data}
\usage{
se1.3M(endpoint = URL_hsds(), svrtype = "hsds", dsetname = "/newassay001")
}
\arguments{
\item{endpoint}{endpoint URL of remote server}

\item{svrtype}{type of server, must be either 'hsds' or 'h5serv'}

\item{dsetname}{complete internal path to dataset in H5 file}
}
\value{
RESTfulSummarizedExperiment
single-cell RNA-seq data from TENx, 1.3mm neurons

SummarizedExperiment instance
}
\description{
Convenience functions using HSDS server to extract tenx neurons full or subset data
}
\note{
se1.3M provides access to the full 1.3 million neurons
with features in their order as given in the original HDF5
}
