% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{se100k}
\alias{se100k}
\title{convenience function using 2017 Channing server to extract 100k tenx neurons subset}
\usage{
se100k(url = "http://54.174.163.77:5000", tag = "tenx_100k_sorted")
}
\arguments{
\item{url}{server URL}

\item{tag}{string giving the internal dataset name}
}
\description{
convenience function using 2017 Channing server to extract 100k tenx neurons subset
}
\examples{
ss = se100k()
# get a set of genes from Tasic et al. 2016 Nature Neuroscience
tc = tasicCortex()
adultCort = tc$GENEID
# subset
csums = apply(assay(ss[adultCort,1:500]),1,sum)
names(csums) = tc$SYMBOL
csums
}
