% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\name{seByTumor_2016}
\alias{seByTumor_2016}
\title{given a BigQueryConnection to the 2016 ISB TCGA bigtables, obtain a SummarizedExperiment 'shell' rowData and colData}
\usage{
seByTumor_2016(tumorCode = "LUAD", assayTblName = "mRNA_UNC_HiSeq_RSEM",
  rdColsToKeep = c("original_gene_symbol", "HGNC_gene_symbol", "gene_id",
  "Study"), bqConn)
}
\arguments{
\item{tumorCode}{one of the concise TCGA codes in a character string -- not checked, defaults to "LUAD", lung adenocarcinoma}

\item{assayTblName}{the name of the assay whose annotation will be used as rowData}

\item{rdColsToKeep}{columns of assay table to use in rowData component}

\item{bqConn}{instance of BigQueryConnection from bigrquery}
}
\description{
given a BigQueryConnection to the 2016 ISB TCGA bigtables, obtain a SummarizedExperiment 'shell' rowData and colData
}
