% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\name{seByTumor_GDC}
\alias{seByTumor_GDC}
\title{given a BigQueryConnection to the 2017 GDC-oriented ISB TCGA bigtables, obtain a SummarizedExperiment 'shell' rowData and colData}
\usage{
seByTumor_GDC(tumorCode = "LUAD",
  assayTblName = "RNAseq_Gene_Expression_UNC_RSEM",
  rdColsToKeep = c("original_gene_symbol", "HGNC_gene_symbol", "gene_id"),
  bqConnClinical, bqConnAssay)
}
\arguments{
\item{tumorCode}{one of the concise TCGA codes in a character string -- not checked, defaults to "LUAD", lung adenocarcinoma}

\item{assayTblName}{the name of the assay whose annotation will be used as rowData}

\item{rdColsToKeep}{columns of assay table to use in rowData component}

\item{bqConnClinical}{instance of BigQueryConnection from bigrquery, for access to clinical metadata -- current expectation is that the BigQuery dataset is named "TCGA_bioclin_v0" and has a table called "Clinical"}

\item{bqConnAssay}{instance of BigQueryConnection from bigrquery -- current expectation is that the BigQuery dataset is named "TCGA_hg19_data_v0"}
}
\description{
given a BigQueryConnection to the 2017 GDC-oriented ISB TCGA bigtables, obtain a SummarizedExperiment 'shell' rowData and colData
}
\examples{
\dontrun{
require(bigrquery)
# provide a valid billing code in the following and authenticate as needed
# clinQ = DBI::dbConnect(dbi_driver(), project = "isb-cgc", 
#       dataset = "TCGA_bioclin_v0", billing = "cgc-xx-xxxx")
# assayQ = DBI::dbConnect(dbi_driver(), project = "isb-cgc", 
#       dataset = "TCGA_hg19_data_v0", billing = "cgc-xx-xxxx")
 myexpShell = seByTumor_GDC( bqConnClinical=clinQ,
      bqConnAssay=assayQ)
 nrow(myexpShell) == 20531
 ncol(myexpShell) == 522
}
}
