% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\name{isplit}
\alias{isplit}
\alias{sproc}
\title{isplit converts a numeric vector into a list of sequences for compact reexpression}
\usage{
isplit(x)

sproc(spl)
}
\arguments{
\item{x}{a numeric vector (should be integers)}

\item{spl}{output of isplit}
}
\description{
isplit converts a numeric vector into a list of sequences for compact reexpression

sproc massages output of isplit into HDF5 select candidates
}
\note{
Very preliminary implementation.
}
\examples{
inds = c(1:10, seq(25,50,2), seq(200,150,-2))
sproc(isplit(inds))
}
