% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st400k}
\alias{st400k}
\title{range-sorted 400k cell subset of 1.3 million cells from 10x genomics}
\format{RangedSummarizedExperiment instance}
\source{
\url{https://community.10xgenomics.com/t5/10x-Blog/Our-1-3-million-single-cell-dataset-is-ready-to-download/ba-p/276}
}
\usage{
st400k
}
\description{
range-sorted 400k cell subset of 1.3 million cells from 10x genomics
}
\note{
used github.com/mtmorgan/TENxGenomics package to obtain 
SummarizedExperiment, then added range information, sorted within
chromosome, and saved shell for use with HDF5 server
}
\examples{
data(st400k)
st400k
}
\keyword{datasets}
