% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{targets}
\alias{targets}
\title{provide the full URLs for link members}
\usage{
targets(h5linkset, index)
}
\arguments{
\item{h5linkset}{instance of H5S_linkset}

\item{index}{numeric index into link vector}
}
\description{
provide the full URLs for link members
}
