% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{tasicCortex}
\alias{tasicCortex}
\title{A set of mouse cortex marker genes.}
\usage{
tasicCortex()
}
\value{
data.frame with columns SYMBOL, GENEID
}
\description{
A set of mouse cortex marker genes.
}
\note{
\url{http://www.nature.com/doifinder/10.1038/nn.4216}, Fig 1C
}
\examples{
head(tasicCortex())
}
