% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tasicST6}
\alias{tasicST6}
\title{tasicST6: supplemental table from Tasic et al. 2016}
\format{DataFrame}
\source{
\url{http://www.nature.com/doifinder/10.1038/nn.4216}
}
\usage{
tasicST6
}
\description{
tasicST6: supplemental table from Tasic et al. 2016
}
\examples{
data(tasicST6)
tasicST6
}
\keyword{datasets}
