% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tenx_100k_sorted}
\alias{tenx_100k_sorted}
\title{tenx_100k_sorted: metadata RangedSummarizedExperiment shell for 100k cells from 10x genomics mouse 1.3 million cell dataset}
\format{RangedSummarizedExperiment instance}
\source{
\url{https://community.10xgenomics.com/t5/10x-Blog/Our-1-3-million-single-cell-dataset-is-ready-to-download/ba-p/276}
}
\usage{
tenx_100k_sorted
}
\description{
tenx_100k_sorted: metadata RangedSummarizedExperiment shell for 100k cells from 10x genomics mouse 1.3 million cell dataset
}
\note{
used github.com/mtmorgan/TENxGenomics package to obtain 
SummarizedExperiment, then added range information, sorted within
chromosome, and saved shell for use with HDF5 server
}
\keyword{datasets}
