% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\docType{methods}
\name{assayNames,BQSummarizedExperiment-method}
\alias{assayNames,BQSummarizedExperiment-method}
\title{Placeholder for assay name extractor for a BQSummarizedExperiment instance.}
\usage{
\S4method{assayNames}{BQSummarizedExperiment}(x, ...)
}
\arguments{
\item{x}{instance of BQSummarizedExperiment}

\item{\dots}{not used}
}
\value{
string indicating that assay is served by BigQuery, nameless
}
\description{
Placeholder for assay name extractor for a BQSummarizedExperiment instance.
}
\note{
This function supplies a placeholder for this early
version of a SummarizedExperiment instance to BigQuery.
At present there is only one assay supported; future work will
help to reduce special coding for BigQuery back end.
}
