% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banoH5.R
\name{banoH5}
\alias{banoH5}
\title{interact with banovichSE content via hdf5 server}
\usage{
banoH5(url = "http://54.163.220.201:5000", host = "assays.hdfgroup.org",
  ...)
}
\description{
interact with banovichSE content via hdf5 server
}
