% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\docType{class}
\name{BQSummarizedExperiment-class}
\alias{BQSummarizedExperiment-class}
\title{define a class to use BigQuery data through SummarizedExperiment interface}
\description{
define a class to use BigQuery data through SummarizedExperiment interface
}
