% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seByTumor.R
\name{cgcConn}
\alias{cgcConn}
\title{Simplify connection to a BigQuery dataset for the project "isb-cgc"}
\usage{
cgcConn(dataset = "TCGA_bioclin_v0", project = "isb-cgc",
  billing = Sys.getenv("CGC_BILLING"))
}
\arguments{
\item{dataset}{character string with dataset name}

\item{project}{character string with project name}

\item{billing}{string with billing code}
}
\value{
instance of \code{\link[bigrquery]{BigQueryConnection-class}}
}
\description{
Simplify connection to a BigQuery dataset for the project "isb-cgc"
}
\examples{
cgcConn
require(bigrquery)
# defaults concern new GDC-compliant format
if (nchar(Sys.getenv("CGC_BILLING"))>0) {
    clin = cgcConn()
    dbListTables(clin)
}
}
