% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\docType{methods}
\name{dim,RESTfulSummarizedExperiment-method}
\alias{dim,RESTfulSummarizedExperiment-method}
\title{dimension access for RESTfulSummarizedExperiment}
\usage{
\S4method{dim}{RESTfulSummarizedExperiment}(x)
}
\arguments{
\item{x}{instance of RESTfulSummarizedExperiment}
}
\value{
vector of nrows, ncols
}
\description{
dimension access for RESTfulSummarizedExperiment
}
