% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{gtexTiss}
\alias{gtexTiss}
\title{Convenience function for access to gene-level GTEx tissues, as quantified in recount}
\usage{
gtexTiss(endpoint = URL_hsds(), svrtype = "hsds",
  dsetname = "/assay001")
}
\arguments{
\item{endpoint}{endpoint URL of remote server}

\item{svrtype}{type of server, must be either 'hsds' or 'h5serv'}

\item{dsetname}{complete internal path to dataset in H5 file}
}
\value{
SummarizedExperiment instance
}
\description{
Convenience function for access to gene-level GTEx tissues, as quantified in recount
}
\examples{
gtexTiss()
}
