% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{gtexTiss}
\alias{gtexTiss}
\title{Convenience function for access to gene-level GTEx tissues, as quantified in recount}
\usage{
gtexTiss(url = URL_hsds(), tag = "assay001")
}
\arguments{
\item{url}{ip address/host for HDF5 server}

\item{tag}{name of hdf5 file on server}
}
\value{
SummarizedExperiment instance
}
\description{
Convenience function for access to gene-level GTEx tissues, as quantified in recount
}
\examples{
gtexTiss()
}
