% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\docType{class}
\name{H5S_dataset-class}
\alias{H5S_dataset-class}
\alias{[,H5S_dataset,character,character,ANY-method}
\title{H5S_dataset class representation}
\usage{
\S4method{[}{H5S_dataset,character,character,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{instance of H5S_dataset}

\item{i}{character string usable as select option for first matrix index in HDF5 server value API}

\item{j}{character string usable as select option for second matrix index in HDF5 server value API}

\item{\dots}{unused}

\item{drop}{logical defaults to FALSE}
}
\description{
H5S_dataset class representation

definition of H5S_dataset subsetting operation
}
\section{Slots}{

\describe{
\item{\code{source}}{instance of H5S_source instance}

\item{\code{simpleName}}{character string naming dataset}

\item{\code{shapes}}{list including dimension information}

\item{\code{hrefs}}{DataFrame of hrefs as defined in the API}

\item{\code{allatts}}{list of all attributes}

\item{\code{presel}}{string prepared for select operation in GET}
}}

