% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbCgc.R
\name{isbCgcDatasets}
\alias{isbCgcDatasets}
\title{vector of dataset names in isb-cgc project}
\usage{
isbCgcDatasets()
}
\value{
character vector
}
\description{
vector of dataset names in isb-cgc project
}
\examples{
isbCgcDatasets() # might be out of date ... can use list_datasets in bigrquery
}
