% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbCgc.R
\name{isbCgcTables}
\alias{isbCgcTables}
\title{list the tables in a selected dataset}
\usage{
isbCgcTables(dataset = "TCGA_hg19_data_v0", billing = .cgcBilling)
}
\arguments{
\item{dataset}{character string identifying a table in ISB CGC}

\item{billing}{Google BigQuery billing code}
}
\description{
list the tables in a selected dataset
}
