% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\docType{methods}
\name{RESTfulSummarizedExperiment}
\alias{RESTfulSummarizedExperiment}
\alias{assay}
\title{construct RESTfulSummarizedExperiment}
\usage{
RESTfulSummarizedExperiment(se, source)

\S4method{assay}{RESTfulSummarizedExperiment,missing}(x, i, ...)
}
\arguments{
\item{se}{SummarizedExperiment instance, assay component can be empty SimpleList}

\item{source}{instance of H5S_dataset}
}
\description{
construct RESTfulSummarizedExperiment
}
\examples{
bigec2 = H5S_source(serverURL="http://54.174.163.77:5000")
n100k = bigec2[["neurons100k"]]
data(tenx_100k_sorted)
rr = RESTfulSummarizedExperiment(tenx_100k_sorted, n100k)
rr
rr2 = rr[1:4, 1:5] # just modify metadata
rr2
assay(rr2) # extract data
}
