% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\name{RESTfulSummarizedExperiment}
\alias{RESTfulSummarizedExperiment}
\title{construct RESTfulSummarizedExperiment}
\usage{
RESTfulSummarizedExperiment(se, source)
}
\arguments{
\item{se}{SummarizedExperiment instance, assay component can be empty SimpleList}

\item{source}{instance of H5SDatasets}
}
\description{
construct RESTfulSummarizedExperiment
}
\examples{
\dontrun{
bb = banoH5()
data(banoSEMeta)
rr = RESTfulSummarizedExperiment(banoSEMeta, bb)
rr
assay(rr)
}
}
