% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RESTfulSE.R
\name{RESTfulSummarizedExperiment}
\alias{RESTfulSummarizedExperiment}
\title{construct RESTfulSummarizedExperiment}
\usage{
RESTfulSummarizedExperiment(se, source)
}
\arguments{
\item{se}{SummarizedExperiment instance, assay component can be empty SimpleList}

\item{source}{instance of H5SDatasets}
}
\description{
construct RESTfulSummarizedExperiment
}
\examples{
\dontrun{
shi = H5S_source(serverURL="http://170.223.248.164:7248")
lin1 = links(shi,1) 
ds1 = datasetRefs(lin1,1,drop=1:4)
n100k = ds1[["neurons100k"]]
data(tenx_100k_sorted)
rr = RESTfulSummarizedExperiment(tenx_100k_sorted, n100k)
rr
assay(rr[1:10,1:20])
}
}
