% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demos.R
\name{se100k}
\alias{se100k}
\alias{se1.3M}
\title{Convenience functions using HSDS server to extract tenx neurons full or subset data}
\usage{
se100k(endpoint = URL_hsds(), svrtype = "hsds",
  dsetname = "/assay001")

se1.3M(endpoint = URL_hsds(), svrtype = "hsds",
  dsetname = "/newassay001")
}
\arguments{
\item{endpoint}{endpoint URL of remote server}

\item{svrtype}{type of server, must be either 'hsds' or 'h5serv'}

\item{dsetname}{complete internal path to dataset in H5 file}
}
\value{
RESTfulSummarizedExperiment

SummarizedExperiment instance
}
\description{
Convenience functions using HSDS server to extract tenx neurons full or subset data
}
\note{
se1.3M provides access to the full 1.3 million neurons
with features in their order as given in the original HDF5
while se100k provides access to only 100k neurons with
expression features sorted by genomic location
}
\examples{
ss = se100k()
# get a set of genes from Tasic et al. 2016 Nature Neuroscience
tc = tasicCortex()
adultCort = tc$GENEID
# subset
csums = apply(assay(ss[adultCort,1:500]),1,sum)
names(csums) = tc$SYMBOL
csums
}
